+++ internal/browser/browserdomain/browser_executable.go
@@ -1,9 +1,5 @@
 package browserdomain
 
-import (
-	. "github.com/git-town/git-town/v22/pkg/prelude"
-)
-
 const NoBrowser = BrowserExecutable("(none)")
 
 // BrowserExecutable indicates a custom browser to use.
@@ -20,21 +16,4 @@ func (self BrowserExecutable) Get() (executable string, useBrowser bool) { //nol
 
 func (self BrowserExecutable) String() string {
 	return string(self)
-}
-
-func ParseBrowser(value, _ string) (Option[BrowserExecutable], error) {
-	if value == "" {
-		return None[BrowserExecutable](), nil
-	}
-	return Some(BrowserExecutable(value)), nil
-}
-
-func ParseBrowserHas(value string, has bool) (Option[BrowserExecutable], Option[BrowserEnabled], error) {
-	if !has {
-		return None[BrowserExecutable](), None[BrowserEnabled](), nil
-	}
-	if value == "" {
-		return None[BrowserExecutable](), Some(BrowserEnabled(false)), nil
-	}
-	return Some(BrowserExecutable(value)), Some(BrowserEnabled(true)), nil
 }
