+++ internal/browser/browserdomain/browser_test.go
@@ -10,14 +10,18 @@ import (
 
 func TestUseBrowser(t *testing.T) {
 	t.Parallel()
-	tests := map[Option[browserdomain.BrowserExecutable]]bool{
-		None[browserdomain.BrowserExecutable]():         true,
-		Some(browserdomain.BrowserExecutable("")):       false,
-		Some(browserdomain.BrowserExecutable("(none)")): false,
-		Some(browserdomain.BrowserExecutable("chrome")): true,
-	}
-	for give, want := range tests {
-		have := browserdomain.BrowserEnabled(give)
-		must.EqOp(t, want, have)
-	}
+
+	t.Run("ParseBrowserHas", func(t *testing.T) {
+		t.Parallel()
+		tests := map[Option[browserdomain.BrowserExecutable]]bool{
+			None[browserdomain.BrowserExecutable]():         true,
+			Some(browserdomain.BrowserExecutable("")):       false,
+			Some(browserdomain.BrowserExecutable("(none)")): false,
+			Some(browserdomain.BrowserExecutable("chrome")): true,
+		}
+		for give, want := range tests {
+			have := browserdomain.BrowserEnabled(give)
+			must.EqOp(t, want, have)
+		}
+	})
 }
